Enhance [Internal] [Middleware Utils] NewCacheMiddleware
- [+] feat(middleware): create helper.go and refactor utils.go
- [+] Add helper.go file with helper functions for generating UUIDs and signatures based on IP, hashing for signature, and middleware options for key generator and cache skipper.
- [+] Refactor utils.go to use helper functions from helper.go and update NewCacheMiddleware function to support custom configuration options.
